Allow advanced formatting on Goage graph
this adds a new formatter for the text in the center of the graph, centerFormatting. this allows the tooltip formatter to be different then the center text. This also allows a data filter through the centerData property.  the results data is passed through, then sent to the centerFormatting. using this method we can now have multiline text in the center, and full access to all of the data within.